requirements.txt
# Core
fastapi>=0.115.0
uvicorn>=0.38.0
python-dotenv>=1.2.1
pydantic>=2.12.0

# HTTP & Parsing
aiohttp>=3.13.0
feedparser>=6.0.12
beautifulsoup4>=4.14.0
lxml>=6.0.0

# AI Providers (install at least one)
anthropic>=0.75.0       # Recommended - supports prompt caching
openai>=1.60.0          # Alternative provider
google-genai>=1.0.0         # Alternative provider

# Advanced Features (optional)
# Install Playwright for JS rendering: pip install playwright && playwright install chromium
playwright>=1.49.0

# Development
pytest>=9.0.0
pytest-asyncio>=1.3.0
